Move multipartable out into Multipart namespace
This is the last of the top level namespace polluting modules moved out.
The tests are all passing (however, I can't comment on how complete the test coverage is)
